page dynamique infogroup et listgroup
web.xml

+++ projet_xml/src/servlet/DeleteFriendFromGroup.java
@@ -9,7 +9,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import classe.Friend;
 import classe.Group;
 import classe.User;
 
@@ -31,7 +30,8 @@ public class DeleteFriendFromGroup extends HttpServlet {
 		Group currentGroup = (Group) session.getAttribute("currentGroup");
 		User user = (User) session.getAttribute("user");
 		String idFriend = request.getParameter("friend");
-		user.deleteFriendFromGroup(idFriend, currentGroup.getName());
+		if(user.groupExist(currentGroup.getName()) && user.friendExist(idFriend))
+			user.deleteFriendFromGroup(idFriend, currentGroup.getName());
 		this.getServletContext().getRequestDispatcher("/consultInfoGroup.jsp").forward(request, response);
 	}
